# openapi.yaml
openapi: 3.0.3
info:
  title: Bing Wallpapers API
  version: 0.1.0
servers:
  - url: https://api.sonurai.com
paths:
  /:
    get:
      operationId: getRoot
      responses:
        '200':
          description: OK
  /wallpapers:
    get:
      operationId: getWallpapers
      summary: Returns a list of wallpapers
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 24
        - in: query
          name: startAfterDate
          required: false
          schema:
            $ref: '#/components/schemas/Date'
        - in: query
          name: startAfterID
          required: false
          schema:
            $ref: '#/components/schemas/ID'
        - in: query
          name: prev
          required: false
          schema:
            type: integer
            enum:
              - 1
      responses:
        '200':
          description: A list of wallpapers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WallpaperList'
        '404':
          description: Not Found
  /wallpapers/{id}:
    get:
      operationId: getWallpaper
      summary: Returns a wallpaper with the given ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WallpaperWithTags'
        '404':
          description: Not Found
  /wallpapers/tags:
    get:
      operationId: getWallpaperTags
      summary: Returns a list of tags
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: integer
        '404':
          description: Not Found
  /wallpapers/tags/{tag}:
    get:
      operationId: getWallpapersByTag
      summary: Returns a list of wallpapers with the given tag
      parameters:
        - in: path
          name: tag
          required: true
          schema:
            type: string
        - in: query
          name: after
          required: false
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WallpaperList'
        '404':
          description: Not Found
components:
  schemas:
    ID:
      type: string
    Date:
      type: integer
      minimum: 19700101
      maximum: 30000101
    Links:
      type: object
      properties:
        prev:
          type: string
        next:
          type: string
    Wallpaper:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        title:
          type: string
        copyright:
          type: string
        date:
          $ref: '#/components/schemas/Date'
        filename:
          type: string
        market:
          type: string
      required:
        - id
        - title
        - copyright
        - date
        - filename
        - market
    WallpaperList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Wallpaper'
        links:
          $ref: '#/components/schemas/Links'
      required:
        - data
        - links
    WallpaperWithTags:
      allOf:
        - $ref: '#/components/schemas/Wallpaper'
        - type: object
          properties:
            tags:
              type: object
              additionalProperties:
                type: number
                format: float
          required:
            - tags